ADAPT MF compile fixes
Becaeuse people modify the PAL headers which impact all projects.

+++ MicroFrameworkPK_v4_3/DeviceCode/Targets/Native/Krait/DeviceCode/Krait_TIMER/Krait__TIMER_Functions.cpp
@@ -33,7 +33,7 @@ BOOL CPU_Timer_UnInitialize(UINT16 Timer)
 }
 
 #define TICKS_PROXIMITY_FORCE 10
-BOOL CPU_Timer_SetCompare(UINT16 Timer, UINT32 CompareValue)
+BOOL CPU_Timer_SetCompare(UINT16 Timer, UINT64 CompareValue)
 {
 	UINT32 diff;
 	UINT32 now, setCompareValue;
